requirements.txt
streamlit
kalshi-python
beautifulsoup4
requests
pandas
plotly
python-dotenv
cryptography
python-dotenv